Add project.vim, gundo, and a special repo for local mods

+++ Rakefile
@@ -42,7 +42,7 @@ def vim_plugin_task(name, repo=nil)
           dirname  = File.basename(filename, '.tar.gz')
 
           sh "tar zxvf tmp/#{filename}"
-          sh "mv #{dirname} #{dir}"
+          sh "test -r ${dirname} || mv #{dirname} #{dir}"
 
         when /vba(\.gz)?$/
           if filename =~ /gz$/
@@ -85,15 +85,17 @@ def vim_plugin_task(name, repo=nil)
       end
 
       task :install => [:pull] + subdirs do
-        Dir.chdir dir do
-          if File.exists?("Rakefile") and `rake -T` =~ /^rake install/
-            sh "rake install"
-          elsif File.exists?("install.sh")
-            sh "sh install.sh"
-          else
-            subdirs.each do |subdir|
-              if File.exists?(subdir)
-                sh "cp -rf #{subdir}/* #{cwd}/#{subdir}/"
+        if File.exists?(dir) 
+          Dir.chdir dir do
+            if File.exists?("Rakefile") and `rake -T` =~ /^rake install/
+              sh "rake install"
+            elsif File.exists?("install.sh")
+              sh "sh install.sh"
+            else
+              subdirs.each do |subdir|
+                if File.exists?(subdir)
+                  sh "cp -rf #{subdir}/* #{cwd}/#{subdir}/"
+                end
               end
             end
           end
@@ -153,6 +155,24 @@ vim_plugin_task "syntastic",        "git://github.com/scrooloose/syntastic.git"
 vim_plugin_task "puppet",           "git://github.com/ajf/puppet-vim.git"
 vim_plugin_task "scala",            "git://github.com/bdd/vim-scala.git"
 vim_plugin_task "gist-vim",         "git://github.com/mattn/gist-vim.git"
+vim_plugin_task "extradite",        "git://github.com/int3/vim-extradite.git"
+vim_plugin_task "gundo",            "git://github.com/sjl/gundo.vim.git"
+vim_plugin_task "pathogen",         "git://github.com/tpope/vim-pathogen.git"
+
+vim_plugin_task "project",          "http://www.vim.org/scripts/download_script.php?src_id=6273" do
+  # Project.vim wants some unique maps that aren't actually unique and error out.
+  sh "sed 's/<unique>//' < plugin/project.vim > project.tmp && mv project.tmp plugin/project.vim"
+end
+
+vim_plugin_task "local",            "git://github.com/AaronBecker/janus_local.git" do
+  sh "cp tmp/local/*.local ."
+  %w[ vimrc.local gvimrc.local ].each do |file|
+    dest = File.expand_path("~/.#{file}")
+    unless File.exist?(dest)
+      ln_s(File.expand_path("../#{file}", __FILE__), dest)
+    end
+  end
+end
 
 vim_plugin_task "command_t",        "git://github.com/wincent/Command-T.git" do
   sh "find ruby -name '.gitignore' | xargs rm"
